ocr-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pytesseract==0.3.13
Pillow==11.1.0
opencv-python-headless==4.10.0.84
pymysql==1.1.1
python-multipart==0.0.20
jinja2==3.1.5
